--- src/main/java/tgits/sealed/chess/Bishop.java
@@ -1,8 +1,0 @@
-package tgits.sealed.chess;
-
-public final class Bishop extends Piece {
-    @Override
-    public String toString() {
-        return "%s Bishop".formatted(this.color.toString());
-    }
-}
