--- src/main/java/tgits/sealed/chess/King.java
@@ -1,8 +1,0 @@
-package tgits.sealed.chess;
-
-public final class King extends Piece {
-    @Override
-    public String toString() {
-        return "%s King".formatted(this.color.toString());
-    }
-}
