--- src/main/java/tgits/sealed/chess/Knight.java
@@ -1,8 +1,0 @@
-package tgits.sealed.chess;
-
-public final class Knight extends Piece {
-    @Override
-    public String toString() {
-        return "%s Knight".formatted(this.color.toString());
-    }
-}
