--- src/main/java/tgits/sealed/chess/Pawn.java
@@ -1,8 +1,0 @@
-package tgits.sealed.chess;
-
-public final class Pawn extends Piece {
-    @Override
-    public String toString() {
-        return "%s Pawn".formatted(this.color.toString());
-    }
-}
