--- src/main/java/tgits/sealed/chess/Queen.java
@@ -1,8 +1,0 @@
-package tgits.sealed.chess;
-
-public final class Queen extends Piece {
-    @Override
-    public String toString() {
-        return "%s Queen".formatted(this.color.toString());
-    }
-}
