--- src/main/java/tgits/sealed/chess/Rook.java
@@ -1,8 +1,0 @@
-package tgits.sealed.chess;
-
-public final class Rook extends Piece {
-    @Override
-    public String toString() {
-        return "%s Rook".formatted(this.color.toString());
-    }
-}
