+++ src/test/java/tgits/sealed/chess/PieceTest.java
@@ -1,0 +1,4 @@
+package tgits.sealed.chess;
+
+public class PieceTest {
+}
